isolate plot_config and plot_script, only add expressions to script for computation
