+++ lib/client/jsx/components/plots/plot.jsx
@@ -1,23 +1,12 @@
 // Framework libraries.
 import * as React from 'react';
 import { connect } from 'react-redux';
-import { selectPlot, plotData } from '../../selectors/plot_selector';
+import { selectPlot } from '../../selectors/plot_selector';
+import { plotData } from '../../plots/plot_script';
 import { MD5, selectConsignment } from '../../selectors/consignment_selector';
 import { requestConsignments } from '../../actions/manifest_actions';
-// import BarGraph from './plots/bar_plot/bar_graph';
-// import BarPlot from './plotter/plots/bar_plot';
-// import BoxGraph from './plots/box_plot/box_graph';
-// import Histogram from '../plotter/plots/histogram';
-import XYPlot, { XYConfig } from './xy_plot/xy_plot';
-import CategoryPlot, { CategoryConfig } from './category_plot/category_plot';
-// import StackedBarPlot from '../plotter/plots/stacked_bar_plot';
-// import Swarm from '../plotter/plots/swarm';
-// import TimelinePlot from './plots/timeline_plot/timeline_graph';
-
-export const PLOTS = {
-  xy: XYConfig,
-  category: CategoryConfig
-};
+import XYPlot from './xy_plot/xy_plot';
+import CategoryPlot from './category_plot/category_plot';
 
 class Plot extends React.Component {
   componentDidMount() {
